added help messages for all tools and displayed them as one

+++ utils/file.py
@@ -49,8 +49,8 @@ def remove_dirs_except(path, dir_exception_list = {}):
 
     for dir in dir_list:
         target_path = os.path.join(path, dir)
-
-        if dir not in dir_exception_list:
+        
+        if os.path.isdir(target_path) and dir not in dir_exception_list:
             shutil.rmtree(target_path)
